Add auth route and move anything relevant for authentication there
