+++ src/controller/index.ts
@@ -1,1 +1,2 @@
 export { default as user } from "./user"
+export { default as auth } from "./auth"
